console: Remove old header
