+++ pkg/webui/components/breadcrumbs/breadcrumbs.styl
@@ -28,6 +28,5 @@
       display: none
 
   &-container
-    border-normal('bottom')
     box-sizing: border-box
     height: $breadcrumbs-bar-height
