+++ pkg/webui/components/header/index.js
@@ -1,4 +1,4 @@
-// Copyright © 2019 The Things Network Foundation, The Things Industries B.V.
+// Copyright © 2023 The Things Network Foundation, The Things Industries B.V.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,107 +12,88 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import React, { useState, useCallback } from 'react'
+import React, { useRef } from 'react'
 import classnames from 'classnames'
 
-import hamburgerMenuNormal from '@assets/misc/hamburger-menu-normal.svg'
-import hamburgerMenuClose from '@assets/misc/hamburger-menu-close.svg'
-
-import NavigationBar from '@ttn-lw/components/navigation/bar'
-import ProfileDropdown from '@ttn-lw/components/profile-dropdown'
-import MobileMenu from '@ttn-lw/components/mobile-menu'
-import Input from '@ttn-lw/components/input'
+import { Breadcrumbs } from '@ttn-lw/components/breadcrumbs/breadcrumbs'
+import Button from '@ttn-lw/components/button'
+import ProfileDropdown from '@ttn-lw/components/profile-dropdown-v2'
+import Link from '@ttn-lw/components/link'
 
 import PropTypes from '@ttn-lw/lib/prop-types'
 
 import style from './header.styl'
 
 const Header = ({
-  className,
-  dropdownItems,
-  navigationEntries,
-  user,
-  searchable,
+  brandLogo,
   logo,
-  mobileDropdownItems,
-  onLogout,
-  onSearchRequest,
+  breadcrumbs,
+  className,
+  addDropdownItems,
+  starDropdownItems,
+  profileDropdownItems,
+  user,
+  onMenuClick,
   ...rest
 }) => {
-  const isGuest = !Boolean(user)
+  const addRef = useRef(null)
+  const starRef = useRef(null)
 
-  const [mobileMenuOpen, setMobileMenuOpen] = useState(false)
-  const handleMobileMenuClick = useCallback(() => {
-    setMobileMenuOpen(!mobileMenuOpen)
-  }, [mobileMenuOpen])
-
-  const handleMobileMenuItemsClick = useCallback(() => {
-    setMobileMenuOpen(false)
-  }, [])
-
-  const classNames = classnames(className, style.container, {
-    [style.mobileMenuOpen]: mobileMenuOpen,
-  })
-
-  const hamburgerGraphic = mobileMenuOpen ? hamburgerMenuClose : hamburgerMenuNormal
+  // Const isGuest = !Boolean(user)
 
   return (
-    <header {...rest} className={classNames}>
-      <div className={style.bar}>
-        <div className={style.left}>
-          {logo}
-          {!isGuest && <NavigationBar className={style.navList}>{navigationEntries}</NavigationBar>}
-        </div>
-        {!isGuest && (
-          <div className={style.right}>
-            {searchable && <Input icon="search" onEnter={onSearchRequest} />}
-            <ProfileDropdown
-              className={style.profileDropdown}
-              userName={user.name || user.ids.user_id}
-              data-test-id="profile-dropdown"
-              profilePicture={user.profile_picture}
-            >
-              {dropdownItems}
-            </ProfileDropdown>
-            <button onClick={handleMobileMenuClick} className={style.mobileMenuButton}>
-              <div className={style.hamburger}>
-                <img src={hamburgerGraphic} alt="Open Mobile Menu" />
-              </div>
-            </button>
-          </div>
-        )}
+    <header {...rest} className={classnames(className, style.container)}>
+      <div className={classnames('breadcrumbs', 's:d-none')} />
+      <div className="d-none s:d-flex al-center gap-cs-xs">
+        <Button secondary icon="menu" onClick={onMenuClick} />
+        <Link to="/" className="d-flex">
+          <img {...logo} className={style.logo} />
+        </Link>
       </div>
-      {mobileMenuOpen && (
-        <MobileMenu
-          className={style.mobileMenu}
-          onItemsClick={handleMobileMenuItemsClick}
-          onLogout={onLogout}
-          user={user}
+
+      <div className="d-flex al-center gap-cs-xs">
+        <Button secondary icon="add" dropdownItems={addDropdownItems} ref={addRef} />
+        <Button
+          secondary
+          icon="grade"
+          dropdownItems={starDropdownItems}
+          ref={starRef}
+          className="s:d-none"
+        />
+        <Button secondary icon="inbox" />
+        <ProfileDropdown
+          brandLogo={brandLogo}
+          data-test-id="profile-dropdown"
+          profilePicture={user?.profile_picture}
         >
-          {mobileDropdownItems}
-        </MobileMenu>
-      )}
+          {profileDropdownItems}
+        </ProfileDropdown>
+      </div>
     </header>
   )
 }
 
+const imgPropType = PropTypes.shape({
+  src: PropTypes.string.isRequired,
+  alt: PropTypes.string.isRequired,
+})
+
 Header.propTypes = {
+  /** The dropdown items when the add button is clicked. */
+  addDropdownItems: PropTypes.node.isRequired,
+  brandLogo: imgPropType,
+  /** A list of breadcrumb elements. */
+  breadcrumbs: PropTypes.arrayOf(PropTypes.oneOfType([PropTypes.func, PropTypes.element]))
+    .isRequired,
   /** The classname applied to the component. */
   className: PropTypes.string,
-  /** The child node of the dropdown component. */
-  dropdownItems: PropTypes.node,
-  /** The logo component. */
-  logo: PropTypes.node.isRequired,
-  /** The child node of the mobile dropdown. */
-  mobileDropdownItems: PropTypes.node,
-  /** The Child node of the navigation bar. */
-  navigationEntries: PropTypes.node,
-  /** A handler for when the user used the search input. */
-  onLogout: PropTypes.func,
-  /** Handler of the search function. */
-  onSearchRequest: PropTypes.func,
-  /* A flag indicating whether the header has a search input. */
-  searchable: PropTypes.bool,
+  logo: imgPropType.isRequired,
+  /** A handler for when the menu button is clicked. */
+  onMenuClick: PropTypes.func.isRequired,
+  /** The dropdown items when the profile button is clicked. */
+  profileDropdownItems: PropTypes.node.isRequired,
+  /** The dropdown items when the star button is clicked. */
+  starDropdownItems: PropTypes.node.isRequired,
   /**
    * The User object, retrieved from the API. If it is `undefined`, then the
    * guest header is rendered.
@@ -122,13 +103,8 @@ Header.propTypes = {
 
 Header.defaultProps = {
   className: undefined,
-  dropdownItems: undefined,
-  navigationEntries: undefined,
-  mobileDropdownItems: null,
-  onSearchRequest: () => null,
-  onLogout: () => null,
-  searchable: false,
   user: undefined,
+  brandLogo: undefined,
 }
 
 export default Header
